Issue 5 - Adding resizable component
