+++ entities/Assets.ts
@@ -1,0 +1,5 @@
+import { Asset as Base } from "@prisma/client";
+
+export interface Asset extends Base {
+  children: Asset[];
+}
