+++ package.json
@@ -15,6 +15,7 @@
     "lodash": "^4.17.21",
     "next": "14.2.3",
     "prisma": "^5.14.0",
+    "re-resizable": "^6.9.17",
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.2.1"
